Refactor code base to packages
