--- src/main/java/EmptyTodoException.java
@@ -1,5 +1,0 @@
-public class EmptyTodoException extends Exception {
-    public EmptyTodoException(String message) {
-        super(message);
-    }
-}
