--- src/main/java/InstructionType.java
@@ -1,11 +1,0 @@
-public enum InstructionType {
-    LIST,
-    MARK,
-    UNMARK,
-    DELETE,
-    TODO,
-    EVENT,
-    DEADLINE,
-    INVALID,
-    BYE
-}
