+++ src/main/java/duke/command/AddEventCommand.java
@@ -1,3 +1,12 @@
+package duke.command;
+
+import duke.common.Messages;
+import duke.storage.Storage;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.ui.TextUi;
+
 import java.time.LocalDateTime;
 
 public class AddEventCommand extends Command {
@@ -12,7 +21,7 @@ public class AddEventCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
+    public void execute(TaskList tasks, TextUi ui, Storage storage) {
         Task task = new Event(title, eventAt);
         tasks.addTask(task);
         ui.showExecutionMessage(Messages.MESSAGE_ADD_EVENT, task.toString(), tasks.getSize());
