+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.common.Messages;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.TextUi;
+
 public class DeleteCommand extends Command {
     private static final boolean IS_EXIT = false;
     private final int taskId;
@@ -8,7 +16,7 @@ public class DeleteCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
+    public void execute(TaskList tasks, TextUi ui, Storage storage) {
         try {
             tasks.deleteTask(taskId);
             ui.showExecutionMessage(Messages.MESSAGE_DELETE);
