+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.common.Messages;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.TextUi;
+
 public class ExitCommand extends Command {
     private static final boolean IS_EXIT = true;
 
@@ -6,7 +14,7 @@ public class ExitCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+    public void execute(TaskList tasks, TextUi ui, Storage storage) throws DukeException {
         storage.saveAllTasks(tasks);
         ui.showExecutionMessage(Messages.MESSAGE_GOOD_BYE);
     }
