+++ src/main/java/duke/command/ListCommand.java
@@ -1,4 +1,9 @@
-import java.util.ArrayList;
+package duke.command;
+
+import duke.DukeException;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.TextUi;
 
 public class ListCommand extends Command {
     private static final boolean IS_EXIT = false;
@@ -8,7 +13,7 @@ public class ListCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
+    public void execute(TaskList tasks, TextUi ui, Storage storage) {
         try {
             for (int i = 0; i < tasks.getSize(); i++) {
                 ui.showToUser((i + 1) + ". " + tasks.getTask(i));
