+++ src/main/java/duke/task/Deadline.java
@@ -1,6 +1,7 @@
+package duke.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 
 public class Deadline extends Task{
     private final LocalDateTime dueBy;
