+++ src/main/java/duke/task/Event.java
@@ -1,6 +1,7 @@
+package duke.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 
 public class Event extends Task{
     private final LocalDateTime eventAt;
