+++ text-ui-test/EXPECTED.TXT
@@ -13,18 +13,18 @@ Now you have 1 task in the list.
   ===================================
   ===================================
 Got it. I've added this task:
-[D][ ] return book (by:  Sunday)
+[D][ ] return book (by:  2/12/2019 1800)
 Now you have 2 tasks in the list.
   ===================================
   ===================================
 Got it. I've added this task:
-[E][ ] project meeting (at:  Mon 2-4pm)
+[E][ ] project meeting (at:  2/12/2019 1800)
 Now you have 3 tasks in the list.
   ===================================
   ===================================
 1. [T][ ] borrow book
-2. [D][ ] return book (by:  Sunday)
-3. [E][ ] project meeting (at:  Mon 2-4pm)
+2. [D][ ] return book (by:  2/12/2019 1800)
+3. [E][ ] project meeting (at:  2/12/2019 1800)
   ===================================
   ===================================
 Nice! I've marked this task as done:
@@ -32,7 +32,7 @@ Nice! I've marked this task as done:
   ===================================
   ===================================
 Nice! I've marked this task as done:
-  [D][X] return book (by:  Sunday)
+  [D][X] return book (by:  2/12/2019 1800)
   ===================================
   ===================================
 OK, I've marked this task as not done yet:
@@ -40,12 +40,12 @@ OK, I've marked this task as not done yet:
   ===================================
   ===================================
 1. [T][ ] borrow book
-2. [D][X] return book (by:  Sunday)
-3. [E][ ] project meeting (at:  Mon 2-4pm)
+2. [D][X] return book (by:  2/12/2019 1800)
+3. [E][ ] project meeting (at:  2/12/2019 1800)
   ===================================
   ===================================
 Noted. I've removed this task:
-  [E][ ] project meeting (at:  Mon 2-4pm)
+  [E][ ] project meeting (at:  2/12/2019 1800)
   ===================================
   ===================================
 Bye. Hope to see you again soon!
